[474793] StandInException in CleanBuildHandler.access$0 (159)
https://bugs.eclipse.org/bugs/show_bug.cgi?id=474793

+++ plugins/org.eclipse.oomph.manifests/src/org/eclipse/oomph/manifests/handlers/CleanBuildHandler.java
@@ -61,25 +61,27 @@ public class CleanBuildHandler extends AbstractHandler
               for (int i = 0; i < array.length; i++)
               {
                 IProject project = (IProject)array[i];
-
-                try
+                if (project.isAccessible())
                 {
-                  IMarker[] markers = project.findMarkers(null, true, IResource.DEPTH_INFINITE);
-                  for (IMarker marker : markers)
+                  try
                   {
-                    try
+                    IMarker[] markers = project.findMarkers(null, true, IResource.DEPTH_INFINITE);
+                    for (IMarker marker : markers)
                     {
-                      marker.delete();
-                    }
-                    catch (CoreException ex)
-                    {
-                      log(ex);
+                      try
+                      {
+                        marker.delete();
+                      }
+                      catch (CoreException ex)
+                      {
+                        log(ex);
+                      }
                     }
                   }
-                }
-                catch (CoreException ex)
-                {
-                  log(ex);
+                  catch (CoreException ex)
+                  {
+                    log(ex);
+                  }
                 }
               }
             }
